+++ changelog.md
@@ -23,6 +23,7 @@ v4.0.7
 * Added Guillotine enchantment, to help you get a head
 * Added a Cryotreatment recipe to turn magma cream into slime balls
 * Fixed a bug where Consecration fields caused suffocation
+* Added Shear spell payload, to help you keep things neatly trimmed
 
 v4.0.6
 ------